Fix kubeseal download, add KinD as a tool
Adds KinD for download via arkade get kind
